Fixed cast for IBM64.
git-svn-id: file:///mnt/fast/tau2git/svnroot/trunk/tau2@4232 a8e31b63-e6ca-409b-a01b-2eec1b506579

+++ src/TraceInput/TAU_tf_writer.cpp
@@ -301,7 +301,7 @@ extern "C" {
     Ttf_fileT *tFile = (Ttf_fileT*)file;
 
     if (tFile->forWriting == false) {
-      return (int)Ttf_CloseFile(tFile);
+      return (int)((long)Ttf_CloseFile(tFile));
     }
 
     for (NidTidMapT::iterator it = tFile->NidTidMap->begin(); it != tFile->NidTidMap->end(); ++it) {
